separated kvdb into 3 crates: kvdb, kvdb-memorydb && kvdb-rocksdb, #6693
