+++ ethcore/light/Cargo.toml
@@ -40,6 +40,8 @@ stats = { path = "../../util/stats" }
 hash = { path = "../../util/hash" }
 triehash = { path = "../../util/triehash" }
 kvdb = { path = "../../util/kvdb" }
+kvdb-rocksdb = { path = "../../util/kvdb-rocksdb" }
+kvdb-memorydb = { path = "../../util/kvdb-memorydb" }
 
 [features]
 default = []
