+++ ethcore/node_filter/Cargo.toml
@@ -11,10 +11,12 @@ ethcore = { path = ".."}
 ethcore-util = { path = "../../util" }
 ethcore-bigint = { path = "../../util/bigint" }
 ethcore-bytes = { path = "../../util/bytes" }
-ethcore-io = { path = "../../util/io" }
 ethcore-network = { path = "../../util/network" }
-kvdb = { path = "../../util/kvdb" }
 native-contracts = { path = "../native_contracts" }
 futures = "0.1"
 log = "0.3"
 parking_lot = "0.4"
+
+[dev-dependencies]
+kvdb-memorydb = { path = "../../util/kvdb-memorydb" }
+ethcore-io = { path = "../../util/io" }
