+++ ethcore/src/client/config.rs
@@ -21,7 +21,7 @@ use std::fmt::{Display, Formatter, Error as FmtError};
 use mode::Mode as IpcMode;
 use verification::{VerifierType, QueueConfig};
 use util::journaldb;
-use kvdb::CompactionProfile;
+use kvdb_rocksdb::CompactionProfile;
 
 pub use std::time::Duration;
 pub use blockchain::Config as BlockChainConfig;
