+++ ethcore/src/lib.rs
@@ -113,6 +113,8 @@ extern crate ansi_term;
 extern crate semantic_version;
 extern crate unexpected;
 extern crate kvdb;
+extern crate kvdb_rocksdb;
+extern crate kvdb_memorydb;
 extern crate util_error;
 extern crate snappy;
 extern crate migration;
