+++ ethcore/src/snapshot/service.rs
@@ -40,7 +40,7 @@ use parking_lot::{Mutex, RwLock, RwLockReadGuard};
 use util_error::UtilError;
 use bytes::Bytes;
 use util::journaldb::Algorithm;
-use kvdb::{Database, DatabaseConfig};
+use kvdb_rocksdb::{Database, DatabaseConfig};
 use snappy;
 
 /// Helper for removing directories in case of error.
@@ -682,7 +682,7 @@ mod tests {
 	#[test]
 	fn cannot_finish_with_invalid_chunks() {
 		use bigint::hash::H256;
-		use kvdb::DatabaseConfig;
+		use kvdb_rocksdb::DatabaseConfig;
 
 		let spec = get_test_spec();
 		let dir = RandomTempPath::new();
