+++ ethcore/src/snapshot/tests/proof_of_authority.rs
@@ -31,7 +31,7 @@ use tests::helpers;
 use transaction::{Transaction, Action, SignedTransaction};
 
 use util::Address;
-use kvdb;
+use kvdb_memorydb;
 
 const PASS: &'static str = "";
 const TRANSITION_BLOCK_1: usize = 2; // block at which the contract becomes activated.
@@ -238,7 +238,7 @@ fn fixed_to_contract_only() {
 	assert_eq!(client.chain_info().best_block_number, 11);
 	let reader = snapshot_helpers::snap(&*client);
 
-	let new_db = kvdb::in_memory(::db::NUM_COLUMNS.unwrap_or(0));
+	let new_db = kvdb_memorydb::in_memory(::db::NUM_COLUMNS.unwrap_or(0));
 	let spec = spec_fixed_to_contract();
 
 	// ensure fresh engine's step matches.
@@ -270,7 +270,7 @@ fn fixed_to_contract_to_contract() {
 
 	assert_eq!(client.chain_info().best_block_number, 16);
 	let reader = snapshot_helpers::snap(&*client);
-	let new_db = kvdb::in_memory(::db::NUM_COLUMNS.unwrap_or(0));
+	let new_db = kvdb_memorydb::in_memory(::db::NUM_COLUMNS.unwrap_or(0));
 	let spec = spec_fixed_to_contract();
 
 	for _ in 0..16 { spec.engine.step() }
