+++ ethcore/src/snapshot/tests/state.rs
@@ -27,7 +27,7 @@ use error::Error;
 use rand::{XorShiftRng, SeedableRng};
 use bigint::hash::H256;
 use util::journaldb::{self, Algorithm};
-use kvdb::{Database, DatabaseConfig};
+use kvdb_rocksdb::{Database, DatabaseConfig};
 use memorydb::MemoryDB;
 use parking_lot::Mutex;
 use devtools::RandomTempPath;
