+++ ethcore/src/tests/helpers.rs
@@ -232,7 +232,7 @@ pub fn get_test_client_with_blocks(blocks: Vec<Bytes>) -> Arc<Client> {
 }
 
 fn new_db() -> Arc<::kvdb::KeyValueDB> {
-	Arc::new(::kvdb::in_memory(::db::NUM_COLUMNS.unwrap_or(0)))
+	Arc::new(::kvdb_memorydb::in_memory(::db::NUM_COLUMNS.unwrap_or(0)))
 }
 
 pub fn generate_dummy_blockchain(block_number: u32) -> BlockChain {
