+++ rpc/Cargo.toml
@@ -57,7 +57,6 @@ rlp = { path = "../util/rlp" }
 stats = { path = "../util/stats" }
 vm = { path = "../ethcore/vm" }
 hash = { path = "../util/hash" }
-kvdb = { path = "../util/kvdb" }
 hardware-wallet = { path = "../hw" }
 
 clippy = { version = "0.0.103", optional = true}
@@ -66,6 +65,7 @@ pretty_assertions = "0.1"
 [dev-dependencies]
 macros = { path = "../util/macros" }
 ethcore-network = { path = "../util/network" }
+kvdb-memorydb = { path = "../util/kvdb-memorydb" }
 
 [features]
 dev = ["clippy", "ethcore/dev", "ethcore-util/dev", "ethsync/dev"]
