+++ sync/src/lib.rs
@@ -48,6 +48,7 @@ extern crate ethcore_light as light;
 
 #[cfg(test)] extern crate ethcore_devtools as devtools;
 #[cfg(test)] extern crate ethkey;
+#[cfg(test)] extern crate kvdb_memorydb;
 
 #[macro_use]
 extern crate macros;
