+++ sync/src/tests/helpers.rs
@@ -291,7 +291,7 @@ impl TestNet<EthPeer<EthcoreClient>> {
 		let client = EthcoreClient::new(
 			ClientConfig::default(),
 			&spec,
-			Arc::new(::kvdb::in_memory(::ethcore::db::NUM_COLUMNS.unwrap_or(0))),
+			Arc::new(::kvdb_memorydb::in_memory(::ethcore::db::NUM_COLUMNS.unwrap_or(0))),
 			Arc::new(Miner::with_spec_and_accounts(&spec, accounts)),
 			IoChannel::disconnected(),
 		).unwrap();
