+++ util/Cargo.toml
@@ -35,6 +35,10 @@ memorydb = { path = "memorydb" }
 util-error = { path = "error" }
 kvdb = { path = "kvdb" }
 
+[dev-dependencies]
+kvdb-rocksdb = { path = "kvdb-rocksdb" }
+kvdb-memorydb = { path = "kvdb-memorydb" }
+
 [features]
 default = []
 dev = ["clippy"]
