+++ util/kvdb-memorydb/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "kvdb-memorydb"
+version = "0.1.0"
+authors = ["debris <marek.kotewicz@gmail.com>"]
+
+[dependencies]
+parking_lot = "0.4"
+rlp = { path = "../rlp" }
+kvdb = { path = "../kvdb" }
