+++ util/src/lib.rs
@@ -110,6 +110,11 @@ extern crate patricia_trie as trie;
 extern crate kvdb;
 extern crate util_error as error;
 
+#[cfg(test)]
+extern crate kvdb_memorydb;
+#[cfg(test)]
+extern crate kvdb_rocksdb;
+
 #[macro_use]
 extern crate log as rlog;
